+++ NEWS.md
@@ -20,6 +20,8 @@ Version 0.17.1b2 - dev
   - `source_block_reference` - the immediate source block reference (`INSERT`), 
     which created the virtual entity, otherwise ``None``
 - NEW: `ezdxf.tools.text_size` module to measure `TEXT` and `MTEXT` entity dimensions
+- CHANGE: `--ltype` arguments of the `draw` command  to `approximate` and `accurate`
+    to be in sync with the `drawing` add-on configuration. 
 - BUGFIX: invalid XDATA processing in `XData.safe_init()` 
 - BUGFIX: group code 1003 is valid in XDATA section 
 - BUGFIX: fix loading error of `DIMSTYLE` attribute `dimtxsty`